Fix possible loss of data (try 3)

+++ src/interface/widget.c
@@ -95,11 +95,11 @@ void widget_scroll_update_thumbs(rct_window *w, int widget_index)
 		scroll->v_thumb_bottom = min(y, view_size);
 
 		if(scroll->v_thumb_bottom - scroll->v_thumb_top < 20) {
-			float percent = (scroll->v_thumb_bottom*1.0)/view_size;
+			double percent = (scroll->v_thumb_bottom*1.0)/view_size;
 			printf("percent: %f\n", percent);
 
-			scroll->v_thumb_top = (uint16) lroundf(scroll->v_thumb_top - (20*percent));
-			scroll->v_thumb_bottom = (uint16) lroundf(scroll->v_thumb_bottom + (20*(1-percent)));
+			scroll->v_thumb_top = (uint16) lround(scroll->v_thumb_top - (20*percent));
+			scroll->v_thumb_bottom = (uint16) lround(scroll->v_thumb_bottom + (20*(1-percent)));
 		}
 	}
